Refactor authentication and error handling
- Updated requirements.txt with new versions of dependencies.
- Refactored authentication routes and forms into a separate blueprint.
- Removed redundant email and error handling modules.
- Updated templates to use new authentication blueprint routes.
- Deleted unused templates and macros.
- Updated translations.
