+++ web_flask/__init__.py
@@ -21,18 +21,23 @@ if os.getenv("FLASK_ENV") == "development":
 else:
     app.config.from_object(Config)
 
+
 mail = Mail(app)
 moment = Moment(app)
 db = SQLAlchemy(app)  # This is the database engine
 migrate = Migrate(app, db)  # This is the migration engine
 
-from .doctor_bp import doctor_bp
-
+from web_flask.errors import bp as errors_bp
+from web_flask.doctor_bp import doctor_bp
+from web_flask.auth import bp as auth_bp
+app.register_blueprint(errors_bp)
 app.register_blueprint(doctor_bp)
+app.register_blueprint(auth_bp)
+
 
 def get_locale():
-    # return request.accept_languages.best_match(app.config['LANGUAGES'])
-    return 'ar'
+    return request.accept_languages.best_match(app.config['LANGUAGES'])
+
 
 babel = Babel(app, locale_selector=get_locale)
 
@@ -75,7 +80,7 @@ login_manager.init_app(app)
 login_manager.login_view = "login"
 login_manager.login_message = _l("Please log in to access this page.")
 
-from web_flask import routes, errors
+from web_flask import routes
 from models.base_model import BaseModel
 import models as m
