+++ web_flask/auth/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('auth', __name__, url_prefix='/auth', template_folder='templates')
+
+from web_flask.auth import routes
