+++ web_flask/auth/email.py
@@ -19,7 +19,7 @@ def send_email(subject, sender, recipients, text_body, html_body):
 def send_password_reset_email(user):
     token = user.get_reset_password_token()
     send_email(
-        "[Raseel] Reset Your Password",
+        "[Raseel Medical Center] Reset Your Password",
         sender=app.config["MAIL_DEFAULT_SENDER"],
         recipients=[user.email],
         text_body=render_template("email/reset_password.txt", user=user, token=token),
