+++ web_flask/auth/templates/register.html
@@ -1,0 +1,6 @@
+{% extends "centeredForms.html" %}
+{% block options %}
+<div class="mt-3">
+    <p>Already have an account? <a href="{{ url_for('auth.login') }}" class="text-decoration-none">Click To Login!</a></p>
+</div>
+{% endblock options %}
