+++ web_flask/routes.py
@@ -1,45 +1,14 @@
 import models as m
 from flask import render_template, redirect, url_for, request, flash, g
-from flask_login import login_user, current_user, logout_user, login_required
+from flask_login import current_user, login_required
 from urllib.parse import urlsplit
 from web_flask import app, db, get_locale
-from web_flask.email import send_password_reset_email
 from web_flask.forms import (
-    LoginForm,
-    RegistrationForm,
     EditProfileInfo,
-    ResetPasswordForm,
-    ResetPasswordRequestForm,
     VisitorForm
 )
+
 import sqlalchemy as sa
-
-
-@app.route("/logout")
-def logout():
-    logout_user()
-    return redirect(url_for("index"))
-
-
-@app.route("/login", methods=["GET", "POST"])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for("edit_profile"))
-    form = LoginForm()
-    if form.validate_on_submit():
-        patient = db.session.scalar(
-            sa.select(m.Patient).where(m.Patient.email == form.email.data)
-        )
-        if patient is None or not patient.check_password(form.password.data):
-            flash("Login Unsuccessful. Please check email and password",
-                  "danger")
-            return redirect(url_for("login"))
-        login_user(patient, remember=form.remember.data)
-        next_page = request.args.get("next")
-        if not next_page or urlsplit(next_page).netloc != "":
-            next_page = url_for("edit_profile")
-        return redirect(next_page)
-    return render_template("login.html", title="Login - Raseel", form=form)
 
 
 @app.route("/", methods=["GET", "POST"])
@@ -51,25 +20,6 @@ def index():
     return render_template("index.html", title="Home - Raseel", form=form)
 
 
-@app.route("/register", methods=["GET", "POST"])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for("edit_profile"))
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        patient_data = {
-            "name": form.name.data,
-            "email": form.email.data,
-            "contact_number": form.contact_number.data,
-            "password": form.password.data,
-            "birth_date": form.birth_date.data,
-        }
-        patient = m.Patient(**patient_data)
-        db.session.add(patient)
-        db.session.commit()
-        flash("Your account has been created! You are now able to log in", "success")
-        return redirect(url_for("login"))
-    return render_template("register.html", title="Register - Raseel", form=form)
 
 
 @app.route("/about")
@@ -111,38 +61,3 @@ def edit_profile():
         form.birth_date.data = current_user.birth_date
     return render_template("edit_profile.html", title="Edit Profile", form=form)
 
-
-@app.route("/reset_password_request", methods=["GET", "POST"])
-def reset_password_request():
-    if current_user.is_authenticated:
-        return redirect(url_for("edit_profile"))
-    form = ResetPasswordRequestForm()
-    if form.validate_on_submit():
-        patient = db.session.scalar(
-            sa.select(m.Patient).where(m.Patient.email == form.email.data)
-        )
-        if patient:
-            send_password_reset_email(patient)
-        flash(
-            ("An email has been sent with instructions to"
-             " reset your password"),
-            "info",
-        )
-        return redirect(url_for("login"))
-    return render_template("reset_password.html", title="Reset Password", heading="Reset Password", form=form)
-
-
-@app.route("/reset_password/<token>", methods=["GET", "POST"])
-def reset_password(token):
-    if current_user.is_authenticated:
-        return redirect(url_for("edit_profile"))
-    patient = m.Patient.verify_reset_password_token(token)
-    if not patient:
-        return redirect(url_for("index"))
-    form = ResetPasswordForm()
-    if form.validate_on_submit():
-        patient.set_password(form.password.data)
-        db.session.commit()
-        flash("Your password has been reset.", "success")
-        return redirect(url_for("login"))
-    return render_template("reset_password.html", title="Reset Password", heading="Reset Password", form=form)
