+++ web_flask/templates/about.html
@@ -266,4 +266,3 @@
             </div>
         </section>
 {% endblock content %}
-
