+++ web_flask/templates/email/reset_password.html
@@ -3,9 +3,9 @@
     <body>
         <p>Hi {{ user.name }},</p>
         <p>Click the link below to reset your password:</p>
-        <a href="{{ url_for('reset_password', token=token, _external=True) }}">Reset Password</a>
+        <a href="{{ url_for('auth.reset_password', token=token, _external=True) }}">Reset Password</a>
         <p>Alternatively, you can paste the following link in your browser:</p>
-        <p>{{ url_for('reset_password', token=token, _external=True) }}</p>
+        <p>{{ url_for('auth.reset_password', token=token, _external=True) }}</p>
         <p>If you did not make this request, simply ignore this email.</p>
         <p>Regards,</p>
         <p>Raseel Healthcare Team</p>
