+++ web_flask/templates/email/reset_password.txt
@@ -2,7 +2,7 @@ Dear {{ user.username }},
 
 To reset your password, visit the following link:
 
-{{ url_for('reset_password', token=token, _external=True) }}
+{{ url_for('auth.reset_password', token=token, _external=True) }}
 
 If you did not make this request then simply ignore this email and no changes will be made.
