--- web_flask/templates/login.html
@@ -1,11 +1,0 @@
-{% extends "centeredForms.html" %}
-{% block options %}
-    <div class="mt-3">
-    <p class="mb-1">
-        New User? <a href="{{ url_for('register') }}" class="text-decoration-none">Register Here!</a>
-    </p>
-    <p class="mb-0">
-        Forgot Password? <a href="{{ url_for('reset_password_request') }}" class="text-decoration-none">Reset Password</a>
-    </p>
-    </div>
-{% endblock options %}
