--- web_flask/templates/register.html
@@ -1,6 +1,0 @@
-{% extends "centeredForms.html" %}
-{% block options %}
-<div class="mt-3">
-    <p>Already have an account? <a href="{{ url_for('login') }}" class="text-decoration-none">Click To Login!</a></p>
-</div>
-{% endblock options %}
