+++ web_flask/translations/ar/LC_MESSAGES/messages.po
@@ -7,8 +7,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
 "Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: 2024-08-28 22:24+0300\n"
-"PO-Revision-Date: 2024-08-28 22:15+0300\n"
+"POT-Creation-Date: 2024-08-29 14:32+0300\n"
+"PO-Revision-Date: 2024-08-29 14:29+0300\n"
 "Last-Translator: \n"
 "Language: ar\n"
 "Language-Team: ar <LL@li.org>\n"
@@ -168,3 +168,4 @@ msgstr ""
 #, python-format
 msgid "%(time_from_now)s"
 msgstr ""
+
